+++ integrationtests/self/timeout_test.go
@@ -70,7 +70,7 @@ var _ = Describe("Timeout tests", func() {
 
 		go func() {
 			defer GinkgoRecover()
-			sess, err := server.Accept()
+			sess, err := server.Accept(context.Background())
 			Expect(err).ToNot(HaveOccurred())
 			str, err := sess.OpenStream()
 			Expect(err).ToNot(HaveOccurred())
@@ -146,7 +146,7 @@ var _ = Describe("Timeout tests", func() {
 			serverSessionClosed := make(chan struct{})
 			go func() {
 				defer GinkgoRecover()
-				sess, err := server.Accept()
+				sess, err := server.Accept(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				sess.AcceptStream() // blocks until the session is closed
 				close(serverSessionClosed)
@@ -187,7 +187,7 @@ var _ = Describe("Timeout tests", func() {
 			serverSessionClosed := make(chan struct{})
 			go func() {
 				defer GinkgoRecover()
-				sess, err := server.Accept()
+				sess, err := server.Accept(context.Background())
 				Expect(err).ToNot(HaveOccurred())
 				sess.AcceptStream() // blocks until the session is closed
 				close(serverSessionClosed)